+++ packages/orm/angel_orm_mysql/test/custom_expr_test.dart
@@ -9,7 +9,7 @@ import 'models/custom_expr.dart';
 void main() {
   Logger.root.level = Level.ALL; // defaults to Level.INFO
   Logger.root.onRecord.listen((record) {
-    print('${record.level.name}: ${record.time}: ${record.message}');
+    print('${record.loggerName}: ${record.time}: ${record.message}');
   });
 
   late MySQLConnection conn;
@@ -36,6 +36,9 @@ void main() {
 
   tearDown(() async {
     await dropTables(runner);
+    if (conn.connected) {
+      await conn.close();
+    }
   });
 
   test('fetches correct result', () async {