+++ packages/orm/angel_orm_mysql/test/edge_case_test.dart
@@ -10,7 +10,7 @@ import 'models/unorthodox.dart';
 void main() {
   Logger.root.level = Level.ALL; // defaults to Level.INFO
   Logger.root.onRecord.listen((record) {
-    print('${record.level.name}: ${record.time}: ${record.message}');
+    print('${record.loggerName}: ${record.time}: ${record.message}');
   });
 
   late MySQLConnection conn;
@@ -34,6 +34,9 @@ void main() {
 
   tearDown(() async {
     await dropTables(runner);
+    if (conn.connected) {
+      await conn.close();
+    }
   });
 
   test('can create object with no id', () async {