+++ packages/orm/angel_orm_mysql/test/enum_test.dart
@@ -6,7 +6,7 @@ import 'models/has_car.dart';
 void main() async {
   Logger.root.level = Level.ALL; // defaults to Level.INFO
   Logger.root.onRecord.listen((record) {
-    print('${record.level.name}: ${record.time}: ${record.message}');
+    print('${record.loggerName}: ${record.time}: ${record.message}');
   });
 
   /// See https://github.com/dart-backend/angel/pull/98