+++ packages/orm/angel_orm_mysql/test/has_many_test.dart
@@ -9,7 +9,7 @@ import 'models/tree.dart';
 void main() {
   Logger.root.level = Level.ALL; // defaults to Level.INFO
   Logger.root.onRecord.listen((record) {
-    print('${record.level.name}: ${record.time}: ${record.message}');
+    print('${record.loggerName}: ${record.time}: ${record.message}');
   });
 
   late MySQLConnection conn;
@@ -28,7 +28,12 @@ void main() {
     treeId = int.parse(appleTree!.id!);
   });
 
-  tearDown(() async => await dropTables(runner));
+  tearDown(() async {
+    await dropTables(runner);
+    if (conn.connected) {
+      await conn.close();
+    }
+  });
 
   test('list is empty if there is nothing', () {
     expect(appleTree!.rings, 10);